remove: undocumented commands that do not work.
