+++ src/actions/action-zoomiso-output-settings.ts
@@ -6,18 +6,13 @@ import { createCommand, sendActionCommand } from './action-utils.js'
 export enum ActionIdZoomISOOutputSettings {
 	setOutputCount = 'set_Output_Count',
 	enableOutput = 'enable_Output',
+	disableOutput = 'disable_Output',
+	setOutputMode = 'set_Output_Mode',
+	setOutputName = 'set_Output_Name',
 	setVideoLossMode = 'set_VideoLoss_Mode',
+	addOutput = 'add_Output',
 	deleteOutput = 'delete_Output',
 	setAudioMode = 'set_Audio_Mode',
-	addOutput = 'add_Output',
-	disableOutput = 'disable_Output',
-	setAudioGainReduction = 'set_AudioGain_Reduction',
-	setOutputSelection = 'set_Output_Selection',
-	setAudioSelection = 'set_Audio_Selection',
-	setOutputEmbeddedAudio = 'set_Output_Embedded_Audio',
-	setOutputName = 'set_Output_Name',
-	setOutputMode = 'set_Output_Mode',
-	setOutputType = 'set_Output_Type',
 }
 
 export function GetActionsZoomISOOutputSettings(instance: InstanceBaseExt<ZoomConfig>): {
@@ -99,11 +94,9 @@ export function GetActionsZoomISOOutputSettings(instance: InstanceBaseExt<ZoomCo
 			options: [options.channel, options.audioChannelMode],
 			callback: (action): void => {
 				// type: 'ISO'
-				// instance.log('debug', `setAudioMode: ${JSON.stringify(action.options)}`)
 				const command = createCommand(instance, '/setAudioMode')
 				command.args.push({ type: 'i', value: action.options.number })
 				command.args.push({ type: 's', value: action.options.audioChannelMode })
-				// instance.log('debug', `setAudioMode: ${JSON.stringify(command)}`)
 				const sendToCommand = {
 					id: ActionIdZoomISOOutputSettings.setAudioMode,
 					options: {
@@ -149,85 +142,6 @@ export function GetActionsZoomISOOutputSettings(instance: InstanceBaseExt<ZoomCo
 				sendActionCommand(instance, sendToCommand)
 			},
 		},
-		[ActionIdZoomISOOutputSettings.setAudioGainReduction]: {
-			name: 'set Audio Gain Reduction',
-			options: [options.channel, options.reductionAmount],
-			callback: (action): void => {
-				// type: 'ISO'
-				const command = createCommand(instance, '/setAudioGainReduction')
-				command.args.push({ type: 'i', value: action.options.number })
-				command.args.push({ type: 'i', value: action.options.reductionAmount })
-
-				const sendToCommand = {
-					id: ActionIdZoomISOOutputSettings.setAudioGainReduction,
-					options: {
-						command: command.oscPath,
-						args: command.args,
-					},
-				}
-				sendActionCommand(instance, sendToCommand)
-			},
-		},
-		[ActionIdZoomISOOutputSettings.setOutputSelection]: {
-			name: 'set Output Selection',
-			description: 'For Spotlight (1-9), Gallery (1-49), and Active Screen Share (1-2)',
-			options: [options.output, options.outputSelectionIndex],
-			callback: (action): void => {
-				// type: 'ISO'
-				const command = createCommand(instance, '/setOutputSelection')
-				command.args.push({ type: 'i', value: action.options.output })
-				command.args.push({ type: 'i', value: action.options.outputSelectionIndex })
-
-				const sendToCommand = {
-					id: ActionIdZoomISOOutputSettings.setOutputSelection,
-					options: {
-						command: command.oscPath,
-						args: command.args,
-					},
-				}
-				sendActionCommand(instance, sendToCommand)
-			},
-		},
-		[ActionIdZoomISOOutputSettings.setAudioSelection]: {
-			name: 'set Audio Selection',
-			description: 'This command does not work in ZoomISO',
-			options: [options.channel, options.reductionAmount],
-			callback: (action): void => {
-				// type: 'ISO'
-				const command = createCommand(instance, '/setAudioSelection')
-				command.args.push({ type: 'i', value: action.options.channel })
-				command.args.push({ type: 'i', value: action.options.reductionAmount })
-
-				const sendToCommand = {
-					id: ActionIdZoomISOOutputSettings.setAudioSelection,
-					options: {
-						command: command.oscPath,
-						args: command.args,
-					},
-				}
-				sendActionCommand(instance, sendToCommand)
-			},
-		},
-		[ActionIdZoomISOOutputSettings.setOutputEmbeddedAudio]: {
-			name: 'set Output Embedded Audio',
-			description: 'Embedded Audio on the outputs.  Starts at 0',
-			options: [options.output, options.mode],
-			callback: (action): void => {
-				// type: 'ISO'
-				const command = createCommand(instance, '/setOutputEmbeddedAudio')
-				command.args.push({ type: 'i', value: action.options.output })
-				command.args.push({ type: 'i', value: action.options.mode })
-
-				const sendToCommand = {
-					id: ActionIdZoomISOOutputSettings.setOutputEmbeddedAudio,
-					options: {
-						command: command.oscPath,
-						args: command.args,
-					},
-				}
-				sendActionCommand(instance, sendToCommand)
-			},
-		},
 		[ActionIdZoomISOOutputSettings.setOutputName]: {
 			name: 'set Output Name',
 			options: [options.output, options.name],
@@ -248,7 +162,6 @@ export function GetActionsZoomISOOutputSettings(instance: InstanceBaseExt<ZoomCo
 				sendActionCommand(instance, sendToCommand)
 			},
 		},
-
 		[ActionIdZoomISOOutputSettings.setOutputMode]: {
 			name: 'setOutputMode',
 			options: [
@@ -278,26 +191,6 @@ export function GetActionsZoomISOOutputSettings(instance: InstanceBaseExt<ZoomCo
 				instance.log('debug', `setOutputMode: ${JSON.stringify(command)}`)
 				const sendToCommand = {
 					id: ActionIdZoomISOOutputSettings.setOutputMode,
-					options: {
-						command: command.oscPath,
-						args: command.args,
-					},
-				}
-				sendActionCommand(instance, sendToCommand)
-			},
-		},
-		// Undocumented command that is not finished
-		[ActionIdZoomISOOutputSettings.setOutputType]: {
-			name: 'setOutputType',
-			description: 'This comment is not finished in ZoomISO',
-			options: [options.output, options.name],
-			callback: (action): void => {
-				// type: 'ISO'
-				const command = createCommand(instance, '/setOutputType')
-				command.args.push({ type: 'i', value: action.options.output })
-				command.args.push({ type: 's', value: action.options.name })
-				const sendToCommand = {
-					id: ActionIdZoomISOOutputSettings.setOutputType,
 					options: {
 						command: command.oscPath,
 						args: command.args,
